Explicitly convert float to int
Implicit conversion from float to int is deprecated in PHP 8.1.

+++ wcfsetup/install/files/lib/system/image/adapter/GDImageAdapter.class.php
@@ -261,7 +261,7 @@ class GDImageAdapter implements IImageAdapter, IWebpImageAdapter
             $this->colorData['red'],
             $this->colorData['green'],
             $this->colorData['blue'],
-            (1 - $opacity) * 127
+            (int) (1 - $opacity) * 127
         );
 
         // draw text
